+++ src/components/routes/BrowerRoutes.jsx
@@ -1,0 +1,14 @@
+import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import App from '../../App';
+import store from '../../redux/store';
+
+const Browser = () => (
+  <BrowserRouter>
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </BrowserRouter>
+);
+
+export default Browser;
