+++ src/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import Browser from './components/routes/BrowerRoutes';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <Browser />
+  </React.StrictMode>,
+);
